update all config to add site ba

+++ app/Http/Controllers/InvComputerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Imports\ImportComputer;
 use App\Models\Aduan;
+use App\Models\Department;
 use App\Models\InvComputer;
 use App\Models\UserAll;
 use Carbon\Carbon;
@@ -19,78 +20,148 @@ class InvComputerController extends Controller
 {
     public function index()
     {
-        $dataInventory = InvComputer::with('pengguna')->get();
-        return Inertia::render('Inventory/Komputer/Komputer', ['komputer' => $dataInventory]);
+        if (auth()->user()->role == 'ict_developer' && auth()->user()->site == 'BIB') {
+            $dataInventory = InvComputer::with('pengguna')->get();
+            $site = '';
+        } else if (auth()->user()->role == 'ict_ho' && auth()->user()->site == 'HO' || auth()->user()->role == 'ict_bod' && auth()->user()->site == 'HO') {
+            $dataInventory = InvComputer::with('pengguna')->where('site', null)->orWhere('site', 'HO')->get();
+
+            $site = '';
+        } else {
+            $dataInventory = InvComputer::with('pengguna')->where('site', auth()->user()->site)->get();
+
+            $site = auth()->user()->site;
+        }
+
+        $role = auth()->user()->role;
+
+        $department = Department::pluck('department_name')->map(function ($name) {
+            return ['name' => $name];
+        })->toArray();
+
+        return Inertia::render('Inventory/Komputer/Komputer', ['komputer' => $dataInventory, 'site' => $site, 'role' => $role, 'department' => $department]);
     }
 
     public function create()
     {
-        // start generate code
-        $currentDate = Carbon::tomorrow();
-        $year = $currentDate->format('y');
-        $month = $currentDate->month;
-        $day = $currentDate->day;
-
-        $maxId = InvComputer::max('max_id');
-
-        if (is_null($maxId)) {
-            $maxId = 0;
-        }
-
-        $uniqueString = 'PPAHOCU' . str_pad(($maxId % 10000) + 1, 3, '0', STR_PAD_LEFT);
-        $request['inventory_number'] = $uniqueString;
-        // end generate code
-
-        $pengguna = UserAll::pluck('username')->map(function ($name) {
-            return ['name' => $name];
-        })->toArray();
-
-        return Inertia::render('Inventory/Komputer/KomputerCreate', ['inventoryNumber' => $uniqueString, 'pengguna' => $pengguna]);
+        abort(404, 'page not found');
     }
 
     public function store(Request $request)
     {
-        // dd($request->all());
-        $maxId = InvComputer::max('max_id');
-        if (is_null($maxId) || empty($maxId)) {
-            $maxId = 1;
-        } else {
-            $maxId = $maxId + 1;
-        }
 
         $params = $request->all();
 
-        $documentation_image = $request->file('image');
-        $destinationPath = 'images/';
-        $path_documentation_image = $documentation_image->store('images', 'public');
-        $new_path_documentation_image = $path_documentation_image;
-        $documentation_image->move($destinationPath, $new_path_documentation_image);
+        if ($params['roterx'] == 'index') {
+            // start generate code
+            $currentDate = Carbon::tomorrow();
+            $year = $currentDate->format('y');
+            $month = $currentDate->month;
+            $day = $currentDate->day;
 
-        $aduan_get_data_user = UserAll::where('username', $params['user_alls_id'])->first();
+            $dept = $params['dept'];
 
-        $data = [
-            'max_id' => $maxId,
-            'computer_name' => $params['computer_name'],
-            'computer_code' => $params['computer_code'],
-            'number_asset_ho' => $params['number_asset_ho'],
-            'assets_category' => $params['assets_category'],
-            'spesifikasi' => $params['model'] . ', ' . $params['processor'] . ', ' . $params['hdd'] . ', ' . $params['ssd'] . ', ' . $params['ram'] . ', ' . $params['vga'] . ', ' . $params['warna_komputer'] . ', ' . $params['os_komputer'],
-            'serial_number' => $params['serial_number'],
-            'aplikasi' => $params['aplikasi'],
-            'license' => $params['license'],
-            'ip_address' => $params['ip_address'],
-            'date_of_inventory' => $params['date_of_inventory'],
-            'date_of_deploy' => $params['date_of_deploy'],
-            'location' => $params['location'],
-            'status' => $params['status'],
-            'condition' => $params['condition'],
-            'note' => $params['note'],
-            'link_documentation_asset_image' => url($new_path_documentation_image),
-            'user_alls_id' => $aduan_get_data_user['id'],
-        ];
+            if (auth()->user()->role == 'ict_developer' && auth()->user()->site == 'BIB') {
+                $maxId = InvComputer::where('site', 'HO')->where('dept', $dept)->orderBy('max_id', 'desc')->first();
 
-        InvComputer::create($data);
-        return redirect()->route('komputer.page');
+                if (is_null($maxId)) {
+                    $maxId = 0;
+                } else {
+                    $jumlah_char_dept = strlen($dept);
+
+                    $jumlah = 8+$jumlah_char_dept;
+                    // dd($jumlah);
+                    $noUrut = (int) substr($maxId->computer_code, $jumlah, 3);
+                    $maxId = $noUrut;
+                }
+
+                $uniqueString = 'BIB-PC-'. $dept . '-' . str_pad(($maxId % 10000) + 1, 3, '0', STR_PAD_LEFT);
+            } else if (auth()->user()->role == 'ict_ho' && auth()->user()->site == 'HO' || auth()->user()->role == 'ict_bod' && auth()->user()->site == 'HO') {
+                $maxId = InvComputer::where('site', 'HO')->where('dept', $dept)->orderBy('max_id', 'desc')->first();
+
+                if (is_null($maxId)) {
+                    $maxId = 0;
+                } else {
+                    $jumlah_char_dept = strlen($dept);
+
+                    $jumlah = 7+$jumlah_char_dept;
+                    // dd($jumlah);
+                    $noUrut = (int) substr($maxId->computer_code, $jumlah, 3);
+                    $maxId = $noUrut;
+                }
+
+                $uniqueString = 'HO-PC-'. $dept . '-' . str_pad(($maxId % 10000) + 1, 3, '0', STR_PAD_LEFT);
+            } else {
+                $maxId = InvComputer::where('site', 'BA')->where('dept', $dept)->orderBy('max_id', 'desc')->first();
+                // dd($maxId->computer_code);
+
+                if (is_null($maxId)) {
+                    $maxId = 0;
+                } else {
+                    $jumlah_char_dept = strlen($dept);
+
+                    $jumlah = 7+$jumlah_char_dept;
+                    // dd($jumlah);
+                    $noUrut = (int) substr($maxId->computer_code, $jumlah, 3);
+                    $maxId = $noUrut;
+                }
+
+                $uniqueString = 'BA-PC-'. $dept . '-' . str_pad(($maxId % 10000) + 1, 3, '0', STR_PAD_LEFT);
+            }
+
+            $request['inventory_number'] = $uniqueString;
+            // end generate code
+
+            $pengguna = UserAll::where('site',auth()->user()->site)->pluck('username')->map(function ($name) {
+                return ['name' => $name];
+            })->toArray();
+
+            return Inertia::render('Inventory/Komputer/KomputerCreate', ['inventoryNumber' => $uniqueString, 'pengguna' => $pengguna]);
+        } else {
+            // dd($request->all());
+            $maxId = InvComputer::max('max_id');
+            if (is_null($maxId) || empty($maxId)) {
+                $maxId = 1;
+            } else {
+                $maxId = $maxId + 1;
+            }
+            
+            $documentation_image = $request->file('image');
+            $destinationPath = 'images/';
+            $path_documentation_image = $documentation_image->store('images', 'public');
+            $new_path_documentation_image = $path_documentation_image;
+            $documentation_image->move($destinationPath, $new_path_documentation_image);
+
+            $aduan_get_data_user = UserAll::where('username', $params['user_alls_id'])->first();
+
+            $dept = explode('-', $params['computer_code']);
+
+            $data = [
+                'max_id' => $maxId,
+                'computer_name' => $params['computer_name'],
+                'computer_code' => $params['computer_code'],
+                'number_asset_ho' => $params['number_asset_ho'],
+                'assets_category' => $params['assets_category'],
+                'spesifikasi' => $params['model'] . ', ' . $params['processor'] . ', ' . $params['hdd'] . ', ' . $params['ssd'] . ', ' . $params['ram'] . ', ' . $params['vga'] . ', ' . $params['warna_komputer'] . ', ' . $params['os_komputer'],
+                'serial_number' => $params['serial_number'],
+                'aplikasi' => $params['aplikasi'],
+                'license' => $params['license'],
+                'ip_address' => $params['ip_address'],
+                'date_of_inventory' => $params['date_of_inventory'],
+                'date_of_deploy' => $params['date_of_deploy'],
+                'location' => $params['location'],
+                'status' => $params['status'],
+                'condition' => $params['condition'],
+                'note' => $params['note'],
+                'link_documentation_asset_image' => url($new_path_documentation_image),
+                'user_alls_id' => $aduan_get_data_user['id'],
+                'site' => auth()->user()->site,
+                'dept' => $dept[2]
+            ];
+
+            InvComputer::create($data);
+            return redirect()->route('komputer.page');
+        }
     }
 
     public function uploadCsv(Request $request)
@@ -109,6 +180,10 @@ class InvComputerController extends Controller
     {
         $komputer = InvComputer::find($id);
 
+        if (empty($komputer)) {
+            abort(404, 'Data not found');
+        }
+
         if (!empty($komputer->user_alls_id)) {
             $aduan_get_data_user = UserAll::where('id', $komputer->user_alls_id)->first()->username;
 
@@ -117,7 +192,7 @@ class InvComputerController extends Controller
             $pengguna_selected = array('data tidak ada !');
         }
 
-        $pengguna_all = UserAll::pluck('username')->map(function ($name) {
+        $pengguna_all = UserAll::where('site',auth()->user()->site)->pluck('username')->map(function ($name) {
             return ['name' => $name];
         })->toArray();
 
@@ -148,8 +223,13 @@ class InvComputerController extends Controller
 
     public function detail($id)
     {
-        $komputer = InvComputer::where('computer_code', $id)->first();
-        $aduan = Aduan::where('inventory_number', $id)->get();
+        $komputer = InvComputer::where('id', $id)->first();
+
+        if (empty($komputer)) {
+            abort(404, 'Data not found');
+        }
+
+        $aduan = Aduan::where('inventory_number', $komputer->computer_code)->get();
         return Inertia::render('Inventory/Komputer/KomputerDetail', [
             'komputer' => $komputer,
             'aduan' => $aduan,
@@ -186,6 +266,7 @@ class InvComputerController extends Controller
                 'note' => $request->note,
                 'link_documentation_asset_image' => url($new_path_documentation_image),
                 'user_alls_id' => $aduan_get_data_user['id'],
+                'site' => auth()->user()->site
             ];
         } else {
 
@@ -209,6 +290,7 @@ class InvComputerController extends Controller
                 'condition' => $request->condition,
                 'note' => $request->note,
                 'user_alls_id' => $aduan_get_data_user['id'],
+                'site' => auth()->user()->site
             ];
         }
 
@@ -218,6 +300,9 @@ class InvComputerController extends Controller
     public function destroy($id)
     {
         $komputer = InvComputer::find($id);
+        if (empty($komputer)) {
+            abort(404, 'Data not found');
+        }
         // return response()->json(['ap' => $komputer]);
         $komputer->delete();
         return redirect()->back();
